move NetworkConfigurationService to global.py
